Lots more comments and details; handle already-connected case.

+++ vpn-connect.sh
@@ -8,58 +8,89 @@
 # from Mac OS Keychain where it is stored securely. All handling of secrets is done in memory
 # and pipes for security; no temp files!
 #
+# Because you can have multiple VPN login credentials stored safely in your Keychain, if you
+# use more than one VPN realm this script makes it much faster to connect. Less typing! Quicker!
+# Even if you only ever use one VPN realm this script eliminates the need for mouse actions
+# and speeds up your getting on the VPN.
+#
 #
 # Usage:
 # 	vpn-connect.sh [ -v ] [ -h ] [ -n "name of login script in Keychain" ] [ -d "name of VPN device" ]
 #
 # WARNING: USING "-v" WILL DISPLAY CREDENTIALS ON YOUR SCREEN. Do not use "-v" in a non-private environment.
 #
-# 
+#
 # Requirements:
 #	- running this on a Mac OS X machine
-# 	- Cisco AnyConnect VPN client software
+# 	- Cisco AnyConnect VPN client software (including /opt/cisco/anyconnect/bin/vpn command)
 # 	- a Mac Keychain item (Password or Secure Note) in the correct format - see below
-# 
-# 
-# 
+# 	- the item(s) should be in your default Keychain, which is indicated in boldface in the keychain list
+#
+#
+# The Keychain Access Mac app is used to create / modify / delete Keychain Items.
+# It is found at Applications --> Utilities --> Keychain Access
+# From the command line: "open /Applications/Utilities/Keychain Access.app"
+#
+# Keychain notes:
+# 	https://support.apple.com/kb/PH20093	- Keychain Access overview
+# 	https://support.apple.com/kb/PH20119	- Secure Notes in Keychain
+# 	https://support.apple.com/kb/PH20097	- Keychain troubleshooting
+# 	https://support.apple.com/kb/PH20094	- Keychain and login password
+#
+#
+#
 ########################################################################################################################
 # Keychain item format details
-# 
-# 
+#
+# Each Keychain Item VPN login "script" must contain four elements for Harvard:
+# 	1) your Harvard Key email address
+# 	2) your Harvard Key password
+# 	3) a blank line (which forces Duo Push)
+# 	4) the letter "y" (to accept the connection terms)
+#
+# *** You can use either the "Secure Note Item" format or the "Password Item" format. ***
+# 	You don't have to use both. You can pick the format that is easier for you.
+#
+#
 # Secure Note Item:
 # 	- must be multiple lines, in this order: [ username#realm, password, blank line, "y" ]
-# 
-# Secure Note Example:
+#
+# Secure Note Item Example:
 #-------------------------------------------------------
 # john_harvard@harvard.edu#vpnrealm
 # p4ssw0rd
-# 
+#
 # y
 #-------------------------------------------------------
-# 
-# 
+#
+#
 # Password Item:
 # 	- must be a SINGLE line with "\n" separating each element
 # 	- elements are the same as, and in the same order as, the Secure Note format
-# 
+#
 # Password Item Example:
 #-------------------------------------------------------
 # john_harvard@harvard.edu#vpnrealm\np4ssw0rd\n\ny\n
 #-------------------------------------------------------
-# 
+#
 # Note that there are two "\n" back-to-back in the Password Item, to represent the
 # end of the password line and the blank line.
-# 
+#
+# Also note that if your Harvard Key password contains a backslash followed by a lower-case "n"
+# as in "\n" then you CANNOT use the Password Item format because that is the line delimiter.
+# You will HAVE TO use the Secure Note format in that case.
+#
+#
 ########################################################################################################################
-# 
+#
 # Notes:
 # 	- in Mac OS X 10.13.1, if you rename a Secure Note in the Keychain you might not ever be able to access
 # 		it again from the CLI!! It appears that the 'security' command can only read the "Service" name
 # 		of a Secure Note, but when you rename an existing one the "Service" name DOES NOT change!
 # 		If you want to change the name of an existing Secure Note and still use it with this script
 # 		(or any other script calling "/usr/bin/security") you may have to delete and re-create the item.
-# 
-# 
+#
+#
 # Shout-outs:
 # https://superuser.com/questions/1153273/access-key-pairs-in-a-macos-keychain-from-the-commandline
 # https://superuser.com/questions/649614/connect-using-anyconnect-from-command-line
@@ -69,8 +100,8 @@
 
 ########################################################################################################################
 # Optional: Customize these variables to your own default, if desired.
-vpn_host="vpn.harvard.edu"			# Set a default - replaced by optional arg from command line
-keychain_script_name="Stefan VPN Login Script" 	# Same - set a default for if there's no arg
+vpn_host="vpn.harvard.edu"			# Set a default - replaced by optional arg from command line "-d"
+keychain_script_name="Stefan VPN Login Script" 	# Same - set a default for if there's no arg "-n" given
 ########################################################################################################################
 
 
@@ -81,7 +112,11 @@ export PATH
 
 ########################################################################################################################
 usage() {
-	echo -e "\nUsage: $0 [ -v ] [ -h ] [ -n \"name of login script in Keychain\" ] [ -d \"name of VPN device\" ]"
+	echo -e "\nUsage: ${0} [ -v ] [ -h ] [ -n \"name of login script in Keychain\" ] [ -d \"name of VPN device\" ]"
+	echo -e "\nDefault values: (edit this script to change)"
+	echo -e "Keychain item default name: \"${keychain_script_name}\""
+	echo -e "VPN target default name: \"${vpn_host}\""
+	echo -e "\nFor full usage and requirements, read the comments in this script ${0}"
 	echo -e "\nWARNING: USING \"-v\" WILL DISPLAY CREDENTIALS ON YOUR SCREEN. Do not use \"-v\" in a non-private environment.\n"
 }
 
@@ -120,13 +155,15 @@ for command in security xxd plutil xmllint vpn open ; do
 	fi
 done
 
-# Test first arg for a leading dash. If ARGV[1] doesn't start with a dash, it's invalid.
-if ! echo -- "${1}" | egrep -q -- '^-' ; then
-	>&2 echo -e "\nError: \"${1}\" is not a valid option."
-	>&2 usage
-	exit 1
-fi
 
+# Test first arg for a leading dash. If ARGV[1] doesn't start with a dash, it's invalid.
+if [[ -n "${1}" ]] ; then	# Only if it's set
+	if ! printf "%s\n" "${1}" | egrep -q -- '^-.' ; then	# use printf for portability since '-n' is a valid arg to 'echo'
+		>&2 echo -e "\nError: \"${1}\" is not a valid option."
+		>&2 usage
+		exit 1
+	fi
+fi
 
 # Take the args if they are present
 while getopts ":vhn:d:" theOption ; do
@@ -165,7 +202,7 @@ fi
 # Figure out if the keychain item was a Password or a Secure Note.
 # If it's a Password we have to evaluate the '\n' to get newlines.
 # If it's a Secure Note, it's going to be an encoded PLIST in XML which we'll have to decode!
-if /bin/echo ${raw_keychain} | grep -q '\\n' ; then
+if /bin/echo "${raw_keychain}" | grep -q '\\n' ; then
 
 	# Password item, so we use 'echo -e' to evaluate the '\n' into newlines
 	login_script=$( echo -e "${raw_keychain}" )
@@ -182,8 +219,27 @@ fi
 # Finally, we should now have everything we need!
 
 
+# Check the state of the connection. If we're already connected, ask what to do!
+if vpn state | grep -q "state: Connected" ; then
+	echo -e "\n*** VPN connection appears to be already established! ***"
+	echo "\"/opt/cisco/anyconnect/bin/vpn state\" reports \"state: Connected\""
+	vpn stats | egrep 'Client Address \(IPv4\)|Profile Name' 	# Display the client-side IP address and realm
+	printf "Do you want to end the current connection and continue? [Y/n] "
+	read answer
+	case ${answer} in
+		n*)	echo "OK - exiting this script to preserve your existing connection."
+			exit 0 ;;
+		*)	echo "OK - terminating your existing connection to set up a new one."
+			vpn disco | grep '>>'	# Short for "disconnect" but I have to use the cool abbreviation!! :-)
+						# The 'grep' is to show just the useful bits.
+			sleep 1 ;;		# Just in case the subsystem needs a sec.
+	esac
+fi
+
+
 # If the regular Mac client app is running, it prevents the CLI from connecting.
-# Quit the UI app if it's found.
+# Quit the UI app if it's found. We'll start it again at the end.
+# (It's handy to have the menu-bar-item showing the status.)
 ui_pid=$( pgrep "Cisco AnyConnect Secure Mobility Client" ) && kill $ui_pid
 
 
@@ -222,8 +278,9 @@ fi
 # Display the state of the connection - which should be "state: Connected" from the "vpn" command.
 if vpn state | grep -q "state: Connected" ; then
 	echo "Connected!"
+	vpn stats | egrep 'Client Address \(IPv4\)|Profile Name' 	# Display the client-side IP address and realm
 else
-	echo "VPN appears not connected when running command \"vpn state\". Something may have gone wrong. Try repeating with \"-v\"."
+	echo "VPN appears not connected when running command \"/opt/cisco/anyconnect/bin/vpn state\". Something may have gone wrong. Try repeating with \"-v\"."
 	exit 1
 fi
 
@@ -231,5 +288,5 @@ fi
 # Also it makes it easy to disconnect the VPN connection, because to do that once the UI app is running
 # you simply quit it. Quitting the app terminates the connection even if it was started from CLI.
 open -j -g "/Applications/Cisco/Cisco AnyConnect Secure Mobility Client.app"
-echo "Note: use the usual Mac OS VPN app to disconnect simply by quitting it, or type \"vpn disconnect\"."
+echo "Note: use the usual Mac OS VPN app to disconnect simply by quitting the app, or type \"/opt/cisco/anyconnect/bin/vpn disconnect\"."
